+++ src/algo/killer2.cpp
@@ -1,5 +1,7 @@
 #include "killer2.hpp"
 
+#include <loguru.hpp>
+
 #include "../pempek_assert.hpp"
 
 using namespace std;
@@ -38,7 +40,7 @@ void Killer2::make_decisions(double date,
                             SortableJobOrder::UpdateInformation *update_info,
                             SortableJobOrder::CompareInformation *compare_info)
 {
-    printf("Date: %g. Available machines: %s\n", date, available_machines.to_string_brackets().c_str());
+    LOG_F(1, "Date: %g. Available machines: %s", date, available_machines.to_string_brackets().c_str());
 
     // Let's update available machines
     for (const string & ended_job_id : _jobs_ended_recently)
@@ -49,7 +51,7 @@ void Killer2::make_decisions(double date,
         current_allocations.erase(ended_job_id);
     }
 
-    printf("Date: %g. Available machines: %s\n", date, available_machines.to_string_brackets().c_str());
+    LOG_F(1, "Date: %g. Available machines: %s", date, available_machines.to_string_brackets().c_str());
 
     // Let's handle recently released jobs
     for (const string & new_job_id : _jobs_released_recently)
@@ -93,5 +95,5 @@ void Killer2::make_decisions(double date,
         _queue->remove_job(job);
     }
 
-    printf("Date: %g. Available machines: %s\n", date, available_machines.to_string_brackets().c_str());
+    LOG_F(1, "Date: %g. Available machines: %s", date, available_machines.to_string_brackets().c_str());
 }